pytest.ini to avoid spurious warnings update some docstrings but not all basic_measure_params can be used again error in compatible catalogs of a group function in the process of updating notebook working on reviving tests
