+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+filterwarnings =
+    ignore:.*the imp module is deprecated.*:DeprecationWarning
