+++ tests/test_draw.py
@@ -1,10 +1,9 @@
 import multiprocessing as mp
 import numpy as np
-import pytest
 import btk
 import btk.sampling_functions
 import btk.survey
-from btk.survey import Rubin, HSC
+from btk.survey import Rubin
 
 
 def get_draw_generator(
@@ -93,42 +92,3 @@ def test_multiprocessing():
     np.testing.assert_array_equal(
         parallel_im["isolated_images"], serial_im["isolated_images"]
     )
-
-
-def test_multiresolution():
-    catalog_name = "data/sample_input_catalog.fits"
-
-    np.random.seed(0)
-    stamp_size = 24.0
-    batch_size = 8
-    cpus = 1
-    multiprocessing = False
-    add_noise = True
-
-    catalog = btk.catalog.CatsimCatalog.from_file(catalog_name)
-    sampling_function = btk.sampling_functions.DefaultSampling(stamp_size=stamp_size)
-    draw_generator = btk.draw_blends.WLDGenerator(
-        catalog,
-        sampling_function,
-        [Rubin, HSC],
-        stamp_size=stamp_size,
-        batch_size=batch_size,
-        multiprocessing=multiprocessing,
-        cpus=cpus,
-        add_noise=add_noise,
-        meas_bands=("i", "i"),
-    )
-    draw_output = next(draw_generator)
-
-    assert (
-        "LSST" in draw_output["blend_list"].keys()
-    ), "Both surveys get well defined outputs"
-    assert (
-        "HSC" in draw_output["blend_list"].keys()
-    ), "Both surveys get well defined outputs"
-    assert draw_output["blend_images"]["LSST"][0].shape[0] == int(
-        24.0 / 0.2
-    ), "LSST survey should have a pixel scale of 0.2"
-    assert draw_output["blend_images"]["HSC"][0].shape[0] == int(
-        24.0 / 0.167
-    ), "HSC survey should have a pixel scale of 0.167"
